// front/src/app/services/Api/model/mfaVerifyDto.ts
/**
 * Books Project
 * The nestjs booksproject API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MfaVerifyDto { 
    code: string;
    rememberMe?: boolean;
}