// front/src/app/services/Api/model/userDto.ts
/**
 * Books Project
 * The nestjs booksproject API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDto { 
    /**
     * User ID
     */
    id: number;
    /**
     * User email
     */
    email: string;
    /**
     * User first name
     */
    firstName: string;
    /**
     * User last name
     */
    lastName: string;
    /**
     * User role
     */
    role: UserDtoRoleEnum;
    /**
     * User image
     */
    imgUrl: string;
    /**
     * Whether MFA is enabled
     */
    isMFAEnabled: boolean;
}
export enum UserDtoRoleEnum {
    NUMBER_0 = 0,
    NUMBER_1 = 1
};